reorganize indexer files
