+++ apps/indexer/local.config
@@ -11,6 +11,7 @@
         "from_block": "0",
         "to_block": "latest",
         "address": "0x04379e472430818f2fe26f25eb1fb3f6dfa54250f90ee93bb3f839028cf80a21",
-        "fetch_interval": 5
+        "fetch_interval": 5,
+        "private_key": "0x0300001800000000300000180000000000030000000000003006001800006600"
     }
 }
