+++ apps/indexer/src/ethereum_indexer/mod.rs
@@ -1,0 +1,5 @@
+pub mod client;
+pub use client::EthereumClient;
+
+pub mod indexer;
+pub use indexer::EthereumIndexer;
