--- apps/indexer/src/events/mod.rs
@@ -1,2 +1,0 @@
-pub mod starknet;
-pub mod ethereum;
