--- apps/indexer/src/indexing/mod.rs
@@ -1,3 +1,0 @@
-pub mod starknet;
-pub mod ethereum;
-pub mod config;
