+++ apps/indexer/src/starknet_indexer/mod.rs
@@ -1,0 +1,5 @@
+pub mod client;
+pub use client::StarknetClient;
+
+pub mod indexer;
+pub use indexer::StarknetIndexer;
